Number rule change to numberic (#35)
Number rule change to numberic

+++ src/DataForm/Field/Number.php
@@ -11,7 +11,7 @@ class Number extends Field
 {
     public $type = "number";
     public $clause = "where";
-    public $rule = ["integer"];
+    public $rule = ["numeric"];
 
     public function build()
     {
